+++ src/components/ProductListItem.module.css
@@ -1,0 +1,32 @@
+.img {
+  max-width: 221px;
+  min-width: 221px;
+  max-height: 221px;
+  min-height: 221px;
+  margin-bottom: 16px;
+  border-radius: 16px;
+}
+
+.description {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.name {
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.price {
+  font-weight: 700;
+}
+
+.favoriteCount {
+  display: flex;
+  align-items: center;
+  font-size: 12px;
+  font-weight: 500;
+  color: #4b5563;
+  gap: 4px;
+}
